fix: show 404 NotFound page
