+++ package.json
@@ -17,6 +17,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-icons": "^5.4.0",
+    "react-router-dom": "^7.8.2",
     "react-use": "^17.6.0",
     "sonner": "^2.0.6",
     "styled-components": "^6.1.19"
