+++ src/NotFound.css
@@ -1,0 +1,14 @@
+/* Global cursor rule (keep hidden on rest of the site if needed) */
+* {
+  /* cursor: none !important; */
+}
+
+/* Force cursor visible on 404 page */
+.notfound-cursor-fix {
+  cursor: auto !important; /* shows normal arrow */
+}
+
+/* Make sure buttons show pointer */
+.notfound-cursor-fix button {
+  cursor: pointer !important;
+}
